requirements.txt
streamlit==1.18.1
google-generativeai==0.2.0
sentence-transformers==2.2.2
scikit-learn==1.0.2
numpy==1.21.6
torch==1.12.1
faiss-cpu==1.7.2
